+++ gcip2/__init__.py
@@ -1,6 +1,6 @@
 from gcip2.pipeline import GitlabCiBuilderImpl, PipelineBuilderImpl
 from gcip2.pipeline.jobs.base import Base, BaseLinux, BaseTask
-from gcip2.pipeline_vars import Predefined
+from gcip2.pipeline_vars import Predefined, PredefinedExternalPullRequest, PredefinedMergeRequest
 from gcip2.project_config import ProjectConfig
 from gcip2.vault import SecretsHandler
 
@@ -11,6 +11,8 @@ __all__ = (
     "BaseLinux",
     "BaseTask",
     "Predefined",
+    "PredefinedMergeRequest",
+    "PredefinedExternalPullRequest",
     "SecretsHandler",
     "ProjectConfig",
 )