+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - updated default tasks
 - fixed module structure
 - changed cli script in module `tasks_core` from `gciptask` to `dothat`
+- added all predefined pipeline variables
 
 ### [08.08.26 - 0.1.4]
 - fixed package `tasks_core` - added `__init__.py`